+++ numpy/core/src/multiarray/ctors.c
@@ -26,113 +26,8 @@
 #include "array_assign.h"
 #include "mapping.h" /* for array_item_asarray */
 #include "scalarmathmodule.h" /* for npy_mul_with_overflow_intp */
+#include "alloc.h"
 #include <assert.h>
-
-#define NBUCKETS 1024 /* number of buckets for data*/
-#define NBUCKETS_DIM 16 /* number of buckets for dimensions/strides */
-#define NCACHE 8 /* 1 + number of cache entries per bucket */
-static void * datacache[NBUCKETS][NCACHE];
-static void * dimcache[NBUCKETS_DIM][NCACHE];
-
-/*
- * very simplistic small memory block cache to avoid more expensive libc
- * allocations
- * base function for data cache with 1 byte buckets and dimension cache with
- * sizeof(npy_intp) byte buckets
- */
-static NPY_INLINE void *
-_npy_alloc_cache(npy_uintp nelem, npy_uintp esz, npy_uint msz,
-                 void * (*cache)[NCACHE], void * (*alloc)(size_t))
-{
-    assert((esz == 1 && cache == datacache) ||
-           (esz == sizeof(npy_intp) && cache == dimcache));
-    if (nelem < msz) {
-        /* first entry is used as fill counter */
-        npy_uintp * idx = (npy_uintp *)&(cache[nelem][0]);
-        if (*idx > 0) {
-            return cache[nelem][(*idx)--];
-        }
-    }
-    return alloc(nelem * esz);
-}
-
-/*
- * return pointer p to cache, nelem is number of elements of the cache bucket
- * size (1 or sizeof(npy_intp)) of the block pointed too
- */
-static NPY_INLINE void
-_npy_free_cache(void * p, npy_uintp nelem, npy_uint msz,
-                void * (*cache)[NCACHE], void (*dealloc)(void *))
-{
-    if (p != NULL && nelem < msz) {
-        /* first entry is used as fill counter */
-        npy_uintp * idx = (npy_uintp *)&(cache[nelem][0]);
-        if (*idx < NCACHE - 1) {
-            cache[nelem][++(*idx)] = p;
-            return;
-        }
-    }
-    dealloc(p);
-}
-
-
-/*
- * array data cache, sz is number of bytes to allocate
- */
-static void *
-npy_alloc_cache(npy_uintp sz)
-{
-    return _npy_alloc_cache(sz, 1, NBUCKETS, datacache, &PyDataMem_NEW);
-}
-
-/* zero initialized data, sz is number of bytes to allocate */
-static void *
-npy_alloc_cache_zero(npy_uintp sz)
-{
-    void * p;
-    NPY_BEGIN_THREADS_DEF;
-    if (sz < NBUCKETS) {
-        p = _npy_alloc_cache(sz, 1, NBUCKETS, datacache, &PyDataMem_NEW);
-        memset(p, 0, sz);
-        return p;
-    }
-    NPY_BEGIN_THREADS;
-    p = PyDataMem_NEW_ZEROED(sz, 1);
-    NPY_END_THREADS;
-    return p;
-}
-
-void
-npy_free_cache(void * p, npy_uintp sz)
-{
-    _npy_free_cache(p, sz, NBUCKETS, datacache, &PyDataMem_FREE);
-}
-
-/*
- * dimension/stride cache, uses a different allocator and is always a multiple
- * of npy_intp
- */
-static void *
-npy_alloc_cache_dim(npy_uintp sz)
-{
-    /* dims + strides */
-    if (NPY_UNLIKELY(sz < 2)) {
-        sz = 2;
-    }
-    return _npy_alloc_cache(sz, sizeof(npy_intp), NBUCKETS_DIM, dimcache,
-                            &PyArray_malloc);
-}
-
-void
-npy_free_cache_dim(void * p, npy_uintp sz)
-{
-    /* dims + strides */
-    if (NPY_UNLIKELY(sz < 2)) {
-        sz = 2;
-    }
-    _npy_free_cache(p, sz, NBUCKETS_DIM, dimcache,
-                    &PyArray_free);
-}
 
 /*
  * Reading from a file or a string.
